pyproject.toml
[tool.poetry]
name = "verbiverse"
version = "0.1.0"
description = ""
authors = ["Layton <caolei6767@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13, >=3.9"
langchain-openai = "^0.1.3"
langchain = "^0.1.16"
langchainhub = "^0.1.15"
pypdf = "^4.2.0"
dashscope = "^1.19.0"
pyside6-fluent-widgets = {extras = ["full"], version = "^1.5.6"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
